+++ Adapter/src/main/java/com/angcyo/dsladapter/DslAdapterStatusItem.kt
@@ -9,28 +9,31 @@ package com.angcyo.dsladapter
  */
 open class DslAdapterStatusItem : BaseDslStateItem() {
 
+    companion object {
+        /**正常状态, 切换到内容*/
+        const val ADAPTER_STATUS_NONE = 0
+        /**空数据*/
+        const val ADAPTER_STATUS_EMPTY = 1
+        /**加载中*/
+        const val ADAPTER_STATUS_LOADING = 2
+        /**错误*/
+        const val ADAPTER_STATUS_ERROR = 3
+    }
+
+    /**刷新回调*/
+    var onRefresh: (DslViewHolder) -> Unit = {
+        L.i("[DslAdapterStatusItem] 触发刷新")
+    }
+
+    //是否已经在刷新, 防止重复触发回调
+    var _isRefresh = false
+
     init {
         itemStateLayoutMap[ADAPTER_STATUS_LOADING] = R.layout.base_loading_layout
         itemStateLayoutMap[ADAPTER_STATUS_ERROR] = R.layout.base_error_layout
         itemStateLayoutMap[ADAPTER_STATUS_EMPTY] = R.layout.base_empty_layout
 
         itemState = ADAPTER_STATUS_NONE
-    }
-
-    companion object {
-        /**正常状态, 切换到内容*/
-        const val ADAPTER_STATUS_NONE = -1
-        /**空数据*/
-        const val ADAPTER_STATUS_EMPTY = 0
-        /**加载中*/
-        const val ADAPTER_STATUS_LOADING = 1
-        /**错误*/
-        const val ADAPTER_STATUS_ERROR = 2
-    }
-
-    /**刷新回调*/
-    var onRefresh: (DslViewHolder) -> Unit = {
-        L.i("[DslAdapterStatusItem] 触发刷新")
     }
 
     override fun onItemBind(
@@ -63,7 +66,7 @@ open class DslAdapterStatusItem : BaseDslStateItem() {
     }
 
     /**返回[true] 表示不需要显示情感图, 即显示[Adapter]原本的内容*/
-    open fun isInAdapterStatus() = itemState != ADAPTER_STATUS_NONE
+    open fun isInAdapterStatus() = itemStateEnable && itemState != ADAPTER_STATUS_NONE
 
     open fun _notifyRefresh(itemHolder: DslViewHolder) {
         if (!_isRefresh) {
@@ -71,9 +74,6 @@ open class DslAdapterStatusItem : BaseDslStateItem() {
             itemHolder.post { onRefresh(itemHolder) }
         }
     }
-
-    //是否已经在刷新
-    var _isRefresh = false
 
     override fun _onItemStateChange(old: Int, value: Int) {
         if (old != value && value != ADAPTER_STATUS_LOADING) {