+++ Adapter/src/main/java/com/angcyo/dsladapter/DslLoadMoreItem.kt
@@ -9,27 +9,6 @@ package com.angcyo.dsladapter
  */
 open class DslLoadMoreItem : BaseDslStateItem() {
 
-    init {
-        itemStateLayoutMap[ADAPTER_LOAD_NORMAL] = R.layout.base_loading_layout
-        itemStateLayoutMap[ADAPTER_LOAD_LOADING] = R.layout.base_loading_layout
-        itemStateLayoutMap[ADAPTER_LOAD_NO_MORE] = R.layout.base_no_more_layout
-        itemStateLayoutMap[ADAPTER_LOAD_ERROR] = R.layout.base_error_layout
-        itemStateLayoutMap[ADAPTER_LOAD_RETRY] = R.layout.base_error_layout
-
-        onItemViewDetachedToWindow = {
-            if (itemEnableLoadMore) {
-                //加载失败时, 下次是否还需要加载更多?
-                if (itemState == ADAPTER_LOAD_ERROR) {
-                    itemState =
-                        ADAPTER_LOAD_RETRY
-                }
-            }
-        }
-        thisAreContentsTheSame = { _, _ ->
-            false
-        }
-    }
-
     companion object {
         /**正常状态, 等待加载更多*/
         const val ADAPTER_LOAD_NORMAL = 0
@@ -40,11 +19,11 @@ open class DslLoadMoreItem : BaseDslStateItem() {
         /**加载失败*/
         const val ADAPTER_LOAD_ERROR = 10
         /**加载失败, 自动重试中*/
-        const val ADAPTER_LOAD_RETRY = 11
+        const val _ADAPTER_LOAD_RETRY = 11
     }
 
     /**是否激活加载更多, 默认关闭*/
-    var itemEnableLoadMore = false
+    override var itemStateEnable: Boolean = false
         set(value) {
             field = value
             itemState = ADAPTER_LOAD_NORMAL
@@ -53,6 +32,30 @@ open class DslLoadMoreItem : BaseDslStateItem() {
     /**加载更多回调*/
     var onLoadMore: (DslViewHolder) -> Unit = {
         L.i("[DslLoadMoreItem] 触发加载更多")
+    }
+
+    //是否已经在加载更多
+    var _isLoadMore = false
+
+    init {
+        itemStateLayoutMap[ADAPTER_LOAD_NORMAL] = R.layout.base_loading_layout
+        itemStateLayoutMap[ADAPTER_LOAD_LOADING] = R.layout.base_loading_layout
+        itemStateLayoutMap[ADAPTER_LOAD_NO_MORE] = R.layout.base_no_more_layout
+        itemStateLayoutMap[ADAPTER_LOAD_ERROR] = R.layout.base_error_layout
+        itemStateLayoutMap[_ADAPTER_LOAD_RETRY] = itemStateLayoutMap[ADAPTER_LOAD_ERROR]!!
+
+        onItemViewDetachedToWindow = {
+            if (itemStateEnable) {
+                //加载失败时, 下次是否还需要加载更多?
+                if (itemState == ADAPTER_LOAD_ERROR) {
+                    itemState =
+                        _ADAPTER_LOAD_RETRY
+                }
+            }
+        }
+        thisAreContentsTheSame = { _, _ ->
+            false
+        }
     }
 
     override fun onItemBind(
@@ -67,12 +70,12 @@ open class DslLoadMoreItem : BaseDslStateItem() {
     override fun _onBindStateLayout(itemHolder: DslViewHolder, state: Int) {
         super._onBindStateLayout(itemHolder, state)
 
-        if (itemEnableLoadMore) {
+        if (itemStateEnable) {
             if (itemState == ADAPTER_LOAD_NORMAL || itemState == ADAPTER_LOAD_LOADING) {
                 _notifyLoadMore(itemHolder)
             } else if (itemState == ADAPTER_LOAD_ERROR) {
                 itemHolder.clickItem {
-                    if (itemState == ADAPTER_LOAD_ERROR || itemState == ADAPTER_LOAD_RETRY) {
+                    if (itemState == ADAPTER_LOAD_ERROR || itemState == _ADAPTER_LOAD_RETRY) {
                         //失败的情况下, 点击触发重新加载
                         _notifyLoadMore(itemHolder)
                         updateAdapterItem()
@@ -93,9 +96,6 @@ open class DslLoadMoreItem : BaseDslStateItem() {
             itemHolder.post { onLoadMore(itemHolder) }
         }
     }
-
-    //是否已经在加载更多
-    var _isLoadMore = false
 
     override fun _onItemStateChange(old: Int, value: Int) {
         if (old != value && value != ADAPTER_LOAD_LOADING) {