+++ Sample/src/main/java/com/angcyo/dsladapter/demo/GroupGridDemoActivity.kt
@@ -53,10 +53,10 @@ class GroupGridDemoActivity : BaseRecyclerActivity() {
                         }
 
                         onItemBindOverride = { itemHolder, itemPosition, adapterItem, _ ->
-                            itemHolder.tv(R.id.fold_button).text =
+                            itemHolder.tv(R.id.fold_button)?.text =
                                 if (itemGroupExtend) "折叠 $itemPosition" else "展开 $itemPosition"
 
-                            itemHolder.tv(R.id.text_view).text =
+                            itemHolder.tv(R.id.text_view)?.text =
                                 "分组${i + 1}" +
                                         "    :${itemHolder.adapterPosition}/${itemHolder.layoutPosition}" +
                                         "    :${spanSizeLookup.getSpanIndex(
@@ -120,7 +120,7 @@ class GroupGridDemoActivity : BaseRecyclerActivity() {
                                     }
                                 }
 
-                                itemHolder.tv(R.id.text_view).text =
+                                itemHolder.tv(R.id.text_view)?.text =
                                     "${spanSizeLookup.getSpanIndex(
                                         itemPosition,
                                         spanCount