+++ Sample/src/main/java/com/angcyo/dsladapter/dsl/DslTextItem.kt
@@ -23,6 +23,6 @@ class DslTextItem : DslAdapterItem() {
         adapterItem: DslAdapterItem
     ) {
         super.onItemBind(itemHolder, itemPosition, adapterItem)
-        itemHolder.v<TextView>(R.id.text_view).text = "文本位置:$itemPosition"
+        itemHolder.v<TextView>(R.id.text_view)?.text = "文本位置:$itemPosition"
     }
 }